LayoutTests/http/wpt/webauthn/public-key-credential-get-failure-u2f-silent.https-expected.txt

CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.get' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.get' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.get' within user activated events.

PASS PublicKeyCredential's [[get]] with malformed sign response in a mock hid authenticator.
PASS PublicKeyCredential's [[get]] with no matched allow credentials in a mock hid authenticator.
PASS PublicKeyCredential's [[get]] with no matched allow credentials in a mock hid authenticator. 2
